Minor fixes on duplicated code (#736)
* remove methods that already defined in the NDArrayAdapter


* refactor toString


* remove sparse NDArray


* fix test
