+++ api/src/main/java/ai/djl/BaseModel.java
@@ -268,6 +268,22 @@ public abstract class BaseModel implements Model {
     }
 
     /** {@inheritDoc} */
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder(200);
+        sb.append("Model (\n\tName: ").append(modelName);
+        if (modelDir != null) {
+            sb.append("\n\tModel location: ").append(modelDir.toAbsolutePath());
+        }
+        sb.append("\n\tData Type: ").append(dataType);
+        for (Map.Entry<String, String> entry : properties.entrySet()) {
+            sb.append("\n\t").append(entry.getKey()).append(": ").append(entry.getValue());
+        }
+        sb.append("\n)");
+        return sb.toString();
+    }
+
+    /** {@inheritDoc} */
     @SuppressWarnings("deprecation")
     @Override
     protected void finalize() throws Throwable {
