+++ mxnet/mxnet-engine/src/main/java/ai/djl/mxnet/engine/MxModel.java
@@ -200,20 +200,4 @@ public class MxModel extends BaseModel {
         dataType = paramNDlist.head().getDataType();
         logger.debug("MXNet Model {} ({}) loaded successfully.", paramFile, dataType);
     }
-
-    /** {@inheritDoc} */
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder(200);
-        sb.append("Model (\n\tName: ").append(modelName);
-        if (modelDir != null) {
-            sb.append("\n\tModel location: ").append(modelDir.toAbsolutePath());
-        }
-        sb.append("\n\tData Type: ").append(dataType);
-        for (Map.Entry<String, String> entry : properties.entrySet()) {
-            sb.append("\n\t").append(entry.getKey()).append(": ").append(entry.getValue());
-        }
-        sb.append("\n)");
-        return sb.toString();
-    }
 }
